refactor: remove pricing.models and pricing.service
