+++ apps/erp/app/modules/sales/index.ts
@@ -1,4 +1,3 @@
-export * from "./pricing";
 export * from "./sales.models";
 export * from "./sales.service";
 export * from "./types";
