--- apps/erp/app/modules/sales/pricing/index.ts
@@ -1,3 +1,0 @@
-export * from "./pricing.models";
-export * from "./pricing.service";
-export * from "./types";
