--- apps/erp/app/modules/sales/pricing/ui/index.ts
@@ -1,4 +1,0 @@
-export { default as PriceListView } from "./PriceListView";
-export { PriceTracePopover } from "./PriceTracePopover";
-export { default as PricingRuleForm } from "./PricingRuleForm";
-export { default as PricingRulesTable } from "./PricingRulesTable";
