+++ apps/erp/app/routes/api+/sales.customer-override.ts
@@ -1,7 +1,7 @@
 import { requirePermissions } from "@carbon/auth/auth.server";
 import type { LoaderFunctionArgs } from "react-router";
 import { data } from "react-router";
-import { getCustomerItemPriceOverride } from "~/modules/sales/pricing";
+import { getCustomerItemPriceOverride } from "~/modules/sales";
 
 export async function loader({ request }: LoaderFunctionArgs) {
   const { client, companyId } = await requirePermissions(request, {
