+++ apps/erp/app/routes/x+/sales+/pricing-rules.tsx
@@ -2,8 +2,8 @@ import { requirePermissions } from "@carbon/auth/auth.server";
 import { VStack } from "@carbon/react";
 import type { LoaderFunctionArgs } from "react-router";
 import { Outlet, useLoaderData } from "react-router";
-import { getPricingRules } from "~/modules/sales/pricing";
-import { PricingRulesTable } from "~/modules/sales/pricing/ui";
+import { getPricingRules } from "~/modules/sales";
+import PricingRulesTable from "~/modules/sales/ui/Pricing/PricingRulesTable";
 import type { Handle } from "~/utils/handle";
 import { path } from "~/utils/path";
 import { getGenericQueryFilters } from "~/utils/query";
